requirements.txt
python-vlc==3.0.102
pysher==1.0.1
pyalsaaudio==0.8.4
twilio==6.18.0
google-cloud-texttospeech==0.2.0
# pygobject==3.30.1
requests
ipdb
